edit: ubx_payload.hpp
float -> double

+++ C++/Navio/Common/ubx_payload.hpp
@@ -5,9 +5,9 @@
 
 struct NavPosllhPayload
 {
-  float lon;   // Longitude [deg]
-  float lat;   // Latitude [deg]
-  float hMSL;  // Height above mean sea level [m]
+  double lon;   // Longitude [deg]
+  double lat;   // Latitude [deg]
+  double hMSL;  // Height above mean sea level [m]
 
   friend std::ostream& operator<<(std::ostream& os, const NavPosllhPayload& arg);
 };
@@ -34,21 +34,21 @@ struct NavPvtPayload
   uint8_t min;    // Minute of hour, range 0..59 (UTC)
   uint8_t sec;    // Seconds of minute, range 0..60 (UTC)
 
-  float lon;   // Longitude [deg]
-  float lat;   // Latitude [deg]
-  float hMSL;  // Height above mean sea level [m]
-  float velN;  // NED north veloity [m/s]
-  float velE;  // NED east veloity [m/s]
-  float velD;  // NED down veloity [m/s]
+  double lon;   // Longitude [deg]
+  double lat;   // Latitude [deg]
+  double hMSL;  // Height above mean sea level [m]
+  double velN;  // NED north veloity [m/s]
+  double velE;  // NED east veloity [m/s]
+  double velD;  // NED down veloity [m/s]
 
   friend std::ostream& operator<<(std::ostream& os, const NavPvtPayload& arg);
 };
 
 struct NavVelnedPayload
 {
-  float velN;  // North velocity component [m/s]
-  float velE;  // East velocity component [m/s]
-  float velD;  // Down velocity component [m/s]
+  double velN;  // North velocity component [m/s]
+  double velE;  // East velocity component [m/s]
+  double velD;  // Down velocity component [m/s]
 
   friend std::ostream& operator<<(std::ostream& os, const NavVelnedPayload& arg);
 };
@@ -60,18 +60,18 @@ struct NavTimegpsPayload
 
 struct NavCovPayload
 {
-  float posCovNN;  // Position covariance matrix value p_NN [m^2]
-  float posCovNE;  // Position covariance matrix value p_NE [m^2]
-  float posCovND;  // Position covariance matrix value p_ND [m^2]
-  float posCovEE;  // Position covariance matrix value p_EE [m^2]
-  float posCovED;  // Position covariance matrix value p_ED [m^2]
-  float posCovDD;  // Position covariance matrix value p_DD [m^2]
-  float velCovNN;  // Velocity covariance matrix value v_NN [m^2/s^2]
-  float velCovNE;  // Velocity covariance matrix value v_NE [m^2/s^2]
-  float velCovND;  // Velocity covariance matrix value v_ND [m^2/s^2]
-  float velCovEE;  // Velocity covariance matrix value v_EE [m^2/s^2]
-  float velCovED;  // Velocity covariance matrix value v_ED [m^2/s^2]
-  float velCovDD;  // Velocity covariance matrix value v_DD [m^2/s^2]
+  double posCovNN;  // Position covariance matrix value p_NN [m^2]
+  double posCovNE;  // Position covariance matrix value p_NE [m^2]
+  double posCovND;  // Position covariance matrix value p_ND [m^2]
+  double posCovEE;  // Position covariance matrix value p_EE [m^2]
+  double posCovED;  // Position covariance matrix value p_ED [m^2]
+  double posCovDD;  // Position covariance matrix value p_DD [m^2]
+  double velCovNN;  // Velocity covariance matrix value v_NN [m^2/s^2]
+  double velCovNE;  // Velocity covariance matrix value v_NE [m^2/s^2]
+  double velCovND;  // Velocity covariance matrix value v_ND [m^2/s^2]
+  double velCovEE;  // Velocity covariance matrix value v_EE [m^2/s^2]
+  double velCovED;  // Velocity covariance matrix value v_ED [m^2/s^2]
+  double velCovDD;  // Velocity covariance matrix value v_DD [m^2/s^2]
 
   friend std::ostream& operator<<(std::ostream& os, const NavCovPayload& arg);
 };
